added empty connect()/disconnect() from #118. split profiler into separate class
