+++ src/ExtendedPdo.php
@@ -24,30 +24,54 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
 {
     /**
      *
-     * The current profile information.
-     *
-     * @var array
-     *
-     */
-    protected $profile = [];
-
-    /**
-     *
-     * A query profiler.
-     *
-     * @var ProfilerInterface
-     *
-     */
-    protected $profiler;
-
-    /**
-     *
      * A specialized statement preparer.
      *
      * @var Rebuilder
      *
      */
     protected $rebuilder;
+
+    /**
+     *
+     * Connects to the database and sets PDO attributes.
+     *
+     * @return null
+     *
+     * @throws \PDOException if the connection fails.
+     *
+     */
+    public function connect()
+    {
+        // do nothing: already connected
+    }
+
+    /**
+     *
+     * Explicitly disconnect by unsetting the PDO instance; does not prevent
+     * later reconnection, whether implicit or explicit.
+     *
+     * @return null
+     *
+     * @throws Exception\CannotDisconnect when the PDO instance was injected
+     * for decoration; manage the lifecycle of that PDO instance elsewhere.
+     *
+     */
+    public function disconnect()
+    {
+        throw new Exception\CannotDisconnect();
+    }
+
+    /**
+     *
+     * Is this instance connected to a database?
+     *
+     * @return bool
+     *
+     */
+    public function isConnected()
+    {
+        return true;
+    }
 
     /**
      *
@@ -59,59 +83,6 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
     public function getPdo()
     {
         return $this;
-    }
-
-    /**
-     *
-     * Begins a transaction and turns off autocommit mode.
-     *
-     * @return bool True on success, false on failure.
-     *
-     * @see http://php.net/manual/en/pdo.begintransaction.php
-     *
-     */
-    public function beginTransaction()
-    {
-        $this->beginProfile(__FUNCTION__);
-        $result = parent::beginTransaction();
-        $this->endProfile();
-        return $result;
-    }
-
-    /**
-     *
-     * Commits the existing transaction and restores autocommit mode.
-     *
-     * @return bool True on success, false on failure.
-     *
-     * @see http://php.net/manual/en/pdo.commit.php
-     *
-     */
-    public function commit()
-    {
-        $this->beginProfile(__FUNCTION__);
-        $result = parent::commit();
-        $this->endProfile();
-        return $result;
-    }
-
-    /**
-     *
-     * Executes an SQL statement and returns the number of affected rows.
-     *
-     * @param string $statement The SQL statement to prepare and execute.
-     *
-     * @return int The number of affected rows.
-     *
-     * @see http://php.net/manual/en/pdo.exec.php
-     *
-     */
-    public function exec($statement)
-    {
-        $this->beginProfile(__FUNCTION__);
-        $affected_rows = parent::exec($statement);
-        $this->endProfile($statement);
-        return $affected_rows;
     }
 
     /**
@@ -454,55 +425,6 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
 
     /**
      *
-     * Returns the profiler object.
-     *
-     * @return ProfilerInterface
-     *
-     */
-    public function getProfiler()
-    {
-        return $this->profiler;
-    }
-
-    /**
-     *
-     * Is a transaction currently active?
-     *
-     * @return bool
-     *
-     * @see http://php.net/manual/en/pdo.intransaction.php
-     *
-     */
-    public function inTransaction()
-    {
-        $this->beginProfile(__FUNCTION__);
-        $result = parent::inTransaction();
-        $this->endProfile();
-        return $result;
-    }
-
-    /**
-     *
-     * Returns the last inserted autoincrement sequence value.
-     *
-     * @param string $name The name of the sequence to check; typically needed
-     * only for PostgreSQL, where it takes the form of `<table>_<column>_seq`.
-     *
-     * @return string
-     *
-     * @see http://php.net/manual/en/pdo.lastinsertid.php
-     *
-     */
-    public function lastInsertId($name = null)
-    {
-        $this->beginProfile(__FUNCTION__);
-        $result = parent::lastInsertId($name);
-        $this->endProfile();
-        return $result;
-    }
-
-    /**
-     *
      * Performs a query with bound values and returns the resulting
      * PDOStatement; array values will be passed through `quote()` and their
      * respective placeholders will be replaced in the query string.
@@ -519,31 +441,7 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
     public function perform($statement, array $values = [])
     {
         $sth = $this->prepareWithValues($statement, $values);
-        $this->beginProfile(__FUNCTION__);
         $sth->execute();
-        $this->endProfile($statement, $values);
-        return $sth;
-    }
-
-    /**
-     *
-     * Prepares an SQL statement for execution.
-     *
-     * @param string $statement The SQL statement to prepare for execution.
-     *
-     * @param array $options Set these attributes on the returned
-     * PDOStatement.
-     *
-     * @return PDOStatement
-     *
-     * @see http://php.net/manual/en/pdo.prepare.php
-     *
-     */
-    public function prepare($statement, $options = [])
-    {
-        $this->beginProfile(__FUNCTION__);
-        $sth = parent::prepare($statement, $options);
-        $this->endProfile($statement, $options);
         return $sth;
     }
 
@@ -569,7 +467,6 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
      */
     public function query($statement, $fetch_mode = 0, $fetch_arg1 = null, $fetch_arg2 = null)
     {
-        $this->beginProfile(__FUNCTION__);
         // ouch this is ugly. what we really wanted was a way to do
         // $sth = call_user_func_array(array(parent, 'query'), $args);
         // but could not find a way to do that with PDO::query() multiple function signatures.
@@ -586,7 +483,6 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
                 $sth = parent::query($statement, $fetch_mode, $fetch_arg1, $fetch_arg2);
                 break;
         }
-        $this->endProfile($sth->queryString);
         return $sth;
     }
 
@@ -618,86 +514,6 @@ class ExtendedPdo extends PDO implements ExtendedPdoInterface
             $value[$k] = parent::quote($v, $parameter_type);
         }
         return implode(', ', $value);
-    }
-
-    /**
-     *
-     * Rolls back the current transaction, and restores autocommit mode.
-     *
-     * @return bool True on success, false on failure.
-     *
-     * @see http://php.net/manual/en/pdo.rollback.php
-     *
-     */
-    public function rollBack()
-    {
-        $this->beginProfile(__FUNCTION__);
-        $result = parent::rollBack();
-        $this->endProfile();
-        return $result;
-    }
-
-    /**
-     *
-     * Sets the profiler object.
-     *
-     * @param ProfilerInterface $profiler
-     *
-     * @return null
-     *
-     */
-    public function setProfiler(ProfilerInterface $profiler)
-    {
-        $this->profiler = $profiler;
-    }
-
-    /**
-     *
-     * Begins a profile entry.
-     *
-     * @param string $function The function starting the profile entry.
-     *
-     * @return null
-     *
-     */
-    protected function beginProfile($function)
-    {
-        // if there's no profiler, can't profile
-        if (! $this->profiler) {
-            return;
-        }
-
-        // retain starting profile info
-        $this->profile['time'] = microtime(true);
-        $this->profile['function'] = $function;
-    }
-
-    /**
-     *
-     * Ends and records a profile entry.
-     *
-     * @param string $statement The statement being profiled, if any.
-     *
-     * @param array $values The values bound to the statement, if any.
-     *
-     * @return null
-     *
-     */
-    protected function endProfile($statement = null, array $values = [])
-    {
-        // is there a profiler in place?
-        if ($this->profiler) {
-            // add an entry to the profiler
-            $this->profiler->addProfile(
-                microtime(true) - $this->profile['time'],
-                $this->profile['function'],
-                $statement,
-                $values
-            );
-        }
-
-        // clear the starting profile info
-        $this->profile = [];
     }
 
     /**
